[Decomp][BattleGround] Cohesion split: extract queue to BattleGroundQueue.cpp
